Add application services use case for retrieving user apps
