+++ features/browser/applicationServices/build.gradle
@@ -2,7 +2,21 @@ apply plugin: 'java-library'
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
+
+    implementation project(':features:common:applicationServices')
+    implementation project(':features:browser:domain')
+
+    // RxJava
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.6'
+
+    // Support annotations
+    implementation 'androidx.annotation:annotation:1.1.0'
+
+    // Unit testing
+    testImplementation 'junit:junit:4.12'
+    testImplementation 'org.mockito:mockito-core:2.23.4'
+    testImplementation 'pl.pragmatists:JUnitParams:1.1.1'
 }
 
-sourceCompatibility = "7"
-targetCompatibility = "7"
+sourceCompatibility = "8"
+targetCompatibility = "8"
