+++ features/browser/domain/src/main/java/com/diskin/alon/appsbrowser/browser/domain/MyClass.java
@@ -1,0 +1,4 @@
+package com.diskin.alon.appsbrowser.browser.domain;
+
+public class MyClass {
+}
