+++ features/browser/featureTest/build.gradle
@@ -27,7 +27,7 @@ dependencies {
     def dagger_version = "2.20"
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation project(':features:common')
+    implementation project(':features:common:presentation')
     implementation project(':features:browser:presentation')
 
     implementation 'androidx.appcompat:appcompat:1.0.2'
