+++ features/browser/presentation/build.gradle
@@ -45,7 +45,7 @@ android {
 dependencies {
     def dagger_version = "2.20"
 
-    implementation project(':features:common')
+    implementation project(':features:common:presentation')
     implementation project(':features:browser:applicationServices')
 
     // Support library
