+++ features/common/applicationServices/src/main/java/com/diskin/alon/appsbrowser/common/applicationservices/Mapper.java
@@ -1,0 +1,6 @@
+package com.diskin.alon.appsbrowser.common.applicationservices;
+
+public interface Mapper<S,R> {
+
+    R map(S source);
+}
