+++ features/common/presentation/proguard-rules.pro
