+++ features/common/presentation/src/androidTest/java/com/diskin/alon/appsbrowser/common/ExampleInstrumentedTest.java
