+++ features/common/presentation/src/main/AndroidManifest.xml
