+++ features/common/presentation/src/main/java/com/diskin/alon/appsbrowser/common/ServiceExecutor.java
