--- features/common/src/main/java/com/diskin/alon/appsbrowser/common/UseCase.java
@@ -1,5 +1,0 @@
-package com.diskin.alon.appsbrowser.common;
-
-public interface UseCase<P,R> {
-    R execute(P param);
-}
